added a default of 0 (false) for deleted fields.
added a required (boolean) field for scarab_r_module_attribute, this field
allows the admin who is setting the attributes that are relevant for a project
to set some of the attributes as required to be set for an issue to be
entered in that project (module).

removed setting the pk explicitly in the id_table scripts, it already had
the pk column defined as autoincrement.  mysql was tryiing to start at
0 for me when adding tables without the key, when some were previously added
with explicit keys.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@224 3c7165bb-d2ef-0310-b724-bafd9a86081f