A couple important (global) changes
1.  I changed the arguments on the select box macro in GlobalMacros.vm
It now takes a business object (AttributeValue) and a String field name
that matches a field name in intake. (it is more flexible this way.)

2.  I swapped the order of columns in R_Module_Attribute to match the order of
names in the table name.  I think this is less confusing.


The rest is related to getting the issue entry working.  You can now
enter an issue, though the dedupe page currently shows all issues as
potential dupes.  The search is not implemented.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@278 3c7165bb-d2ef-0310-b724-bafd9a86081f