used torque's alias attribute to get a ScarabUser that uses turbine's user table.
reorganized some of the security sql as it needs to come after the MODULE table
is created.

Added a PERMISSION field to ATTRIBUTE.  This is used to set the PERMISSION
needed for a user to be valid for a UserAttribute.  It is possible that we
may just want to create a specific class for each type of UserAttribute and
implement a getPermission method in UserAttribute.  But I think it might be
generally usefull to declare a permission for some other types of Attributes.

For example only allow user's with EDIT_PRIORITY permission to set the priority
attribute.

john mcnally

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@488 3c7165bb-d2ef-0310-b724-bafd9a86081f