added checking to make sure that values were not null
modified the api so that the result of the getAttributeOptions(boolean)
method is a List instead of an array of objects. this is a cleaner api
that more closely matches what the base objects will represent in the
future. this is a coding convention i want us to follow going forward (as
well as fix in current code as we go along).

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@893 3c7165bb-d2ef-0310-b724-bafd9a86081f