the formatting of dates is in the presentation layer.
instead of having to put a method into each BO, it is better to allow
people to specify this within the templates by using a Format tool. the
format of the date is held as a constant in ScarabConstants.

as we need more formatting tasks, please add them to the Format.java
class.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@906 3c7165bb-d2ef-0310-b724-bafd9a86081f