2-2-0_test.html this is an alternate version of the Adv query screen that uses pop-ups instead of open text entry... if we want to use this style of ui it will need to be made consistent throughout the app. which is not that big of a deal from my perspective (would take two hours).
2-6-0_test.html
this screen shows the query chaining UI that John and I discussed this morning - we should review this one in the meeting today - I would be interested as to what the use cases are (to make sure we have the correct options in place)... John and I were wondering if this was not something we could table for later.

that's it.
B


git-svn-id: http://scarab.tigris.org/svn/scarab@953 3c7165bb-d2ef-0310-b724-bafd9a86081f