changes to query page UI...
couple of important things here. made all the date range form elements pop-up menus... note the "hour" pop-up.

also, personnel search. this is not "hard wired" as we discussed yesterday, but I recognize an enormous oversight in not calling it out in the 2.2.0 query page (and on the saved query page 2.2.1)

Brian


git-svn-id: http://scarab.tigris.org/svn/scarab@1133 3c7165bb-d2ef-0310-b724-bafd9a86081f