added support for specifying a query within a single parameter.  (Basically storing what could be an entire query string under one parameter within the query string.)  This was started as an effort to allow query chaining, though it is not strictly necessary with the current method of just allowing the user to return to the query and alter it, the method leaves room for adding boolean operations joining queries into more complex queries.
Also modified the issue list from a query to be generated from the search
criteria on every request.  So there is no need to store the results on the
server between requests.  (this could be further optimized to only generate
the results needed for one page, but that needs to be examined in a db agnostic
way, which is a difficult problem to solve.)

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@1145 3c7165bb-d2ef-0310-b724-bafd9a86081f