a boat load of changes related to Attributes
you will need to re-create your database.

there is still more cleanup to do, but i need to get this in now so that i
can work on finishing it up. the only editor that works is the global
editor...now i need to go back and finish up the rest of the attribute
editors, but it is easier now that i have things figured out and the code
cleaned up more.

there is tons of commented out code in the Attribute* classes...please
don't remove it yet...

i commented out some fields in intake.xml that didn't seem to be used
right now (according to the errors in the turbine.log file that i got).
john/elicia, please test and see if i screwed something up.

i added a copy of the email that Pier is sending out with bugzilla on
nagoya, it is a neato report format. eventualy it would be cool to do
something like that for scarab.


git-svn-id: http://scarab.tigris.org/svn/scarab@1244 3c7165bb-d2ef-0310-b724-bafd9a86081f