I made changes to Query workflow. If this meets approval, I will do the same for enter issue templates:
User submits a query. If it is a personal query, it is automatically approved. If it is a global (project-wide for all users) query, and the user does not have Query | Approve permission, it must be approved. An email goes out to the project owner to approve or reject it, and on the list page it appears as type "personal (until approved)". If it gets rejected, the type gets set to personal. If it gets approved, the type gets set to global.

When the user visits the QueryList page, they see first all their personal queries, which they have the ability to delete and edit, and then the already-approved global queries for this project. If the user has Item | Delete permission, they can delete the global queries.

Note: this means that once a global query has been approved and gone global, the user who submitted it can no longer delete and edit it.


git-svn-id: http://scarab.tigris.org/svn/scarab@1287 3c7165bb-d2ef-0310-b724-bafd9a86081f