most changes related to searching for issues based on user attributes such as assigned to.
AttributeValues can be chained together when multiple values for an attribute
exist.  This makes it easy to assign multiple attributes to an attribute but one
must be sure to check for chained av's when looping through a list.

also did some cleanup in IssueSearch so that irrelevant sql chunks are not
added when some fields are left blank.

added a Choose... option to the sortby select list, so that every list is not
sorted by summary  value by default.

Since the user selections were multivalued, i had to do some work towards
allowing it, should be straightforward to add for other option attributes as well.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@1294 3c7165bb-d2ef-0310-b724-bafd9a86081f