latest torque latest tomcat (with some fixes i made for classreloading)
i believe i have pretty much finished implementing the
GlobalAttributeEditor with regards to editing AttributeOptions. still need
to add the ability to create a new option, but it seems that the hard
parts are mostly over now.

there is a boat load of business logic behind the AO's and the
ROptionOption's...especially with regards to editing the data via the
webapp. i tried to document it in the code...still need to document it
further.

the caching schemes is a mess. i have some ideas on how we can clean it up
in the future (with events). for now, this is going to have to do.


git-svn-id: http://scarab.tigris.org/svn/scarab@1329 3c7165bb-d2ef-0310-b724-bafd9a86081f