latest tomcat 4.0.1-dev
the beauty of this is that you can now pretty much leave tomcat running
all the time without having to constantly restart it during development.
the classloader should reload the changed classes, which causes a
classloader dump and restart of Turbine.

here is the bug that was fixed:
http://nagoya.apache.org/bugzilla/show_bug.cgi?id=3733

now, we still have the problem where it logs you out, but i'm getting
closer to solving that problem as well...

needless to say, this couldn't have been done with closed source software.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@1341 3c7165bb-d2ef-0310-b724-bafd9a86081f