cleaned up intake.xml...fixed some error messages by setting mapToProperty=""
added Register to intake.xml to handle the registration process

re-wrote the entire registration process to use intake. it is much cleaner
now and the code can be used as a fairly easy example on how to use intake
properly as well as code Turbine applications...

cleaned up the registration process html files to look better UI wise and
use more of the current CSS system...this will of course need to be
updated to the latest CSS when that is ready.

cleaned up javadoc in a few classes

since intake uses the ScarabUser object, it needs to have the Retrievable
methods available and implemented. By default, ScarabUserImpl has those
methods, but other implementations of ScarabUser may not...not we require
it.

added a couple (4) more helper methods to the base actions...should help
remove a lot of duplicate code in the rest of actions (TODO)


git-svn-id: http://scarab.tigris.org/svn/scarab@1455 3c7165bb-d2ef-0310-b724-bafd9a86081f