modified the process of storing the user in the session, so that it is keyed off a query parameter.  This will allow multiple browsers within the same session to enter issues.
moved cleanup code that is to be executed after an issue is entered or the
process is terminated into a method so that the several slightly different
versions could be kept uniform.

removed storing the start point of the issue entry process in the session.
This simple string parameter can be passed in the request parameters.
Added the parameter to $link, so it can be used as a general way to store
a screen that should be saved for several requests.

Issue number:
Obtained from:
Submitted by:
Reviewed by:
Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@1497 3c7165bb-d2ef-0310-b724-bafd9a86081f