first pass at module editor stuff...
includes gary's contributions as well as a boat load of my own changes.

please note the FIXME comment in the ModifyModule.java action...it is a
pretty major issue. if someone could spend some time looking into it, i
would appreciate it...ACL information is getting cached somewhere and it
is driving me nuts trying to figure out where.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@1670 3c7165bb-d2ef-0310-b724-bafd9a86081f