an issue should always have a type.  it also always belongs to a module.  so i modified the constuctors/factory methods to help codify these requirements.
as a result issuetype does not need to be passed to any Issue methods.  I have removed the argument from the java code and have attempted to get it out of the
templates, though it is likely I have missed some in the templates.


Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@1873 3c7165bb-d2ef-0310-b724-bafd9a86081f