created a new class that contains much of the functionality that will be shared across standalone and sourcecast driven implementations.
currently we are using the default structure for the security service's user
entity, so we will need to delegate to the new class.  If we could drive the
structure of the security classes from scarab then we could inherit the
functionality.  But for now, user methods that contain code relevant to
both implementations should follow the model shown in ScarabUserImpl.

The helm implementation should be able to inherit, so it will be simpler.


Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@1952 3c7165bb-d2ef-0310-b724-bafd9a86081f