torque is moving towards requiring the database name to be specified. torque will generate code to create the database and the name is required to use this task.
Added a create-db target that can be used instead of the scripts to
create and populate the database.

the database name is given by the property scarab.database.name.  It can be
changed from the default value of "scarab", which is needed if several
instances of scarab are using the same db server.  Note: if you change the
value of this property, you will need to run the upgrade-fulcrum target.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@2067 3c7165bb-d2ef-0310-b724-bafd9a86081f