this is the start of a transform for bugzilla dtd -> scarab dtd. it is incomplete as it doesn't deal with everything it needs to, but it is a good start. woo hoo!
the .xsl was contributed by the bitonic guys.

i wrote the simple build.xml to use
Ant's <style> task. however (and I hope you guys are reading this), we
should write our own Ant task to replace the use of the <style> task which
we can use to do the transform ourselves (similar to the class you guys
gave me...just make it an ant task instead of calling it from shell
scripts)...that will allow us to have more control over the transform
process as well as do things like pretty print the output (rightnow, it
comes out on one line (this may be something that can be controlled in
the stylesheet though...i'm not sure)...

thanks.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@2186 3c7165bb-d2ef-0310-b724-bafd9a86081f