updates from kevin @ bitonic...
Added in the committed-by element to the issue element.  This allows you
to specify who committed
the issue.  Added the data and created-date elements to the attachment
elements.  This allows for
comments data to be specified as attachments.  When importing from
bugzilla a comment, it is
important to specify the created date because scarab displays the created
date as the date of the
comment.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@2189 3c7165bb-d2ef-0310-b724-bafd9a86081f