Adds multiple selects to Advanced Query, so that you can choose issues with platform is HP or DEC for example.
The macro for the select box was modified to add a multiple flag, so the macro
call in other templates where you should be only selecting one have been
modified to pass false for the flag.

The functionality was already present in the search code.  Just had to make
some small changes to AttributeValue and fix some bugs.

Issue number: 88
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@2738 3c7165bb-d2ef-0310-b724-bafd9a86081f