Fixes issue: http://scarab.tigris.org/issues/show_bug.cgi?id=72
Hitting save and then done creates two artifact types with the same info.
Simple fix is to remove the save button.  There is no reason to return
to this page if the save is successful and the issue type and be edited from
the template returned to after hitting done.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@2954 3c7165bb-d2ef-0310-b724-bafd9a86081f