Removed getIssueList(); now templates call the getCurrentSearchResults() method directly.
Now all the set up for searches happens in getCurrentSearchResults. It searches on the current qu
ery, which is set in the user session every time the user selects a query to execute, or submits
the search form. This method calls getIssueIdsFromUserSearch() which searches on the user attribu
tes if any have been selected.

Issue number: 46, 147


git-svn-id: http://scarab.tigris.org/svn/scarab@3135 3c7165bb-d2ef-0310-b724-bafd9a86081f