was reading a copy of some java magazine and they had an article of a cool *optional* javascript feature that essentially puts focus() on the first form element on a page. i also noticed that google does this as well...so, i decided to add it to a couple of pages to see if people liked it...
note that the javascript needs to appear at the bottom of the page
because it has to execute *after* the form element that it is referring
to is defined. there is a way around this, which is to add a function to
the <body onload="">, but this would require that we also have that
function on each and every page... not really possible here...so, this
will have to do for now.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@3418 3c7165bb-d2ef-0310-b724-bafd9a86081f