A cache whose default implementation is to create one Map per thread. It will be used to cache the results of expensive method calls, so that if the method is called again within the same request cycle, the cached result is returned.
This type of caching is fairly safe to apply globally in cases where
a db query is being made.  In most these cases the method is called
multiple times within the template and the same result is expected
for the second+ calls to the method.  It is possible that there are
cases where an action uses a method which returns a query result and
then modifies the data, so that the query should return a different
result on the next invocation.  As most of the time the method is going
to return a List, the action should just make sure to alter the list
at the same time it is altering the stored data.

A Valve implementation is used to make sure the cache is reset at the
beginning and end of the pipeline.  It is reset at the end as this is
the preferred location to free up resources as soon as possible.  It is
done at the beginning as well, in case an exception prevented cleanup
on a previous request.  The valve is used in the event of thread pooling.
It also provides a useful location to log total reqest processing time.

Testing on ViewIssue.vm has shown a reduction from 1s to 600ms when applying
the cache and it has not been applied everywhere it can be, yet. Results
are likely to be better under load.

Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://scarab.tigris.org/svn/scarab@3552 3c7165bb-d2ef-0310-b724-bafd9a86081f