initial add of a script which will eventually replace the create-mysql-database.sh script
this script allows proper passing in of command line options and will
also support both mysql and postgresql (as well as any future databases)
we need to support.

If you have no username/password, all one has to do to create a mysql
scarab database on localhost is:

    ./create-db.sh mysql

note: this script currently does not support postgresql yet. i will
be working on adding that shortly.

-jon


git-svn-id: http://scarab.tigris.org/svn/scarab@3703 3c7165bb-d2ef-0310-b724-bafd9a86081f