When clicking Cancel in the EditUser.vm and EditRole.vm template, the user returns now to the search results instead of an empty user search. ( See also http://scarab.tigris.org/issues/show_bug.cgi?id=111 )
This is done by storing the last search result into the temporary data of the
user, together with the search parameters (or should this be stored into
the cache instead?). This data can be accessed through some new methods
in the ScarabRequestTool.

I'm not sure this is the best way to do it, but if it's not good, plz restore the
previous version of the modified files, or explain to me how to do it right.

Maarten


git-svn-id: http://scarab.tigris.org/svn/scarab@3739 3c7165bb-d2ef-0310-b724-bafd9a86081f