removed the illegal characters as we do not want to restrict the complexity of queries. a user might want more than a simple list of words.  Instead we pass the query unchanged to lucene.  If lucene cannot parse the query we return a message showing the bad query and the message from lucene on what was the cause.  The user can then edit their query and try again.
For example, a user might want to look for fore* using a wildcard, so we do
not want to disallow the * character.  However lucene does not allow the
wildcard to be the first character, so *ground will give an error.  Instead
of trying to sort out what lucene will and won't accept, it is better left
to the user to learn the system.

An alternate solution to http://scarab.tigris.org/issues/show_bug.cgi?id=177
Also iz# 8496


git-svn-id: http://scarab.tigris.org/svn/scarab@3978 3c7165bb-d2ef-0310-b724-bafd9a86081f