Backed out the changes I made yesterday to the caching of Issue because John was already working on it (sorry John for the overhead).
I don't know the status of this new caching code, but here is (in my point of
view) a possible solution. The main problem is that Issue related things are now
changed outside the knowledge of the Issue. For example adding a comment
happens by creating an Attachment and saving it directly into the database
with calling the Attachment.save() method. This means you will see a lot of
code in ModifyIssue.java that looks like this:

Attachment att = new Attachment();
....  // setting the values of the attachment
att.save();

Because Issue don't know that a new attachment (such as a comment, ...) is
created, it won't update its cache. That's why I think that the code above
should be changed in something like this:

Attachment att = new Attachment();
...  // setting the values of the attachment
issue.addComment(att);  // suppose we want to add a comment
issue.save();

Now, the issue knows a comment is added and it can take action to update
all related caches (such as the caching of the getComments() and
getActivity() methods).

Maarten


git-svn-id: http://scarab.tigris.org/svn/scarab@4058 3c7165bb-d2ef-0310-b724-bafd9a86081f